+++ backend/changelog.md
@@ -1,6 +1,14 @@
+v0.2 :
+- httponly cookie
+- SSR for profile page
+- logout (now available on profile page)
+- password hashing
+- store user id instead of email in cookie
+
 v0.1 :
 - Login & signup page 
 - authentication with cookie
 - dynamic navigation
 - profile page
 - some refactoring 
+